refactor: standardize file and folder naming conventions
